% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_hazard_frailty_type.R
\name{causal.hazard.frailty.type}
\alias{causal.hazard.frailty.type}
\title{Calculation of the causal hazard ratio for different frailty distribution}
\usage{
causal.hazard.frailty.type(
  frailty.type = 1,
  haz1,
  haz0,
  cum.haz1,
  cum.haz0,
  theta
)
}
\arguments{
\item{frailty.type}{}

\item{haz1}{}

\item{haz0}{}

\item{cum.haz1}{}

\item{cum.haz0}{}

\item{theta}{}
}
\value{
a list with 
\itemize{
\item{FOO}{ - BAR}
\item{FOO2}{ - BAR2}
}
}
\description{
This function calculate the causal hazard ratio for the Gamma or the Inverse Gaussian frailty distributions
}
\details{
This function calculates the causal hazard ratio given the hazard rate and the cumulative hazard rate for treatment arm 1 and the hazard rate and the cumulative hazard in treatment arm 0.
The treatment arm 1, will be considered as the "treatment arm" and the treatment arm 0, will be considered as the "control arm".
The hazard ratio will be calculated as the hazard rate in the treatment arm vs. the hazard rate in the control arm.
The hazard rates and the cumulative hazards will be substitutes in the following formulas depending on the chosen frailty distributation.
For the Gamma frailty:
hr<-(haz1/haz0)*exp(theta*(cum.haz1-cum.haz0))
 For the Inverse Gaussian frailty 
 hr<-(haz1/haz0)*((1+theta*cum.haz1)/(1+theta*cum.haz0))
Additional technical details, this function calls \code{print} and also \code{\link{plot}}
}
\examples{
 \dontrun{
 print('hello')
 }
}
\references{
Brill, Barak, Amnon Amir, esting for Differential Abundance in Compositional Counts Data, with Application to Microbiome Studies. arXiv Preprint arXiv:1904.08937.
}
