% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muhaz_by_tau.R
\name{data.by.tau}
\alias{data.by.tau}
\title{Title}
\usage{
data.by.tau(
  tau,
  data,
  frailty.type,
  confounder,
  n.est.grid,
  b.cor,
  end.time,
  min.time,
  kern,
  data.tau = NA
)
}
\arguments{
\item{tau}{}

\item{data}{}

\item{frailty.type}{}

\item{confounder}{}

\item{n.est.grid}{}

\item{b.cor}{}

\item{end.time}{}

\item{min.time}{}

\item{kern}{}

\item{data.tau}{}
}
\value{

}
\description{
Title
}
