% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_for_bootstrap_real_data.R
\name{estimation.for.bootsrap.real.data}
\alias{estimation.for.bootsrap.real.data}
\title{The wrapper function of the bootstrap estimation}
\usage{
estimation.for.bootsrap.real.data(
  data,
  indices,
  frailty.type,
  tau,
  confounder,
  n.est.grid,
  b.cor,
  end.time,
  min.time,
  kern,
  data.tau = NA
)
}
\arguments{
\item{data}{The data that the estimators will be estimated on. In case of RCT, a data with column named "time, treatment, status" needs to be specified. In case of observational studies a firth column named "W" that contains the weights need to be specified also.}

\item{indices}{allows boot to select sample, dont need to be sppecified}

\item{frailty.type}{This function can calculate the causal hazard ratio under two frailty distribution-1=Gamma frailty distribution, 2=Inverse Gaussian distribution. The default value is the Gamma distribution.}

\item{tau}{The wanted Kendall's tau. Under the Gamma frailty the possible values of Kendall's tau are [0,1] and under the inverse Gaussian [0,0.5)}

\item{confounder}{"no"- for RCT setting, or in case a non-weighted estimator wanted to be calculated. "yes"-for observational studies setting, or in evey other case a weighted estimator need to be calculated. The defoult option is "no".}

\item{n.est.grid}{Number of points in the estimation grid, where hazard estimates are computed. Default value is 101.}

\item{b.cor}{Boundary correction type. Possible values are: "none" - no boundary correction "left" - left only correction "both" - left and right corrections Default value is "both". Only the first letter needs to be given (e.g. b.cor="n").}

\item{end.time}{Right bound of the time domain used in analysis.}

\item{min.time}{Left bound of the time domain used in analysis. If missing, min.time is considered 0.}

\item{kern}{Boundary kernel function to be used. Possible values are: "rectangle", "epanechnikov", "biquadratic", "triquadratic". Default value is "epanechnikov". Only the first letter needs to be given (e.g. kern="b").}

\item{data.tau}{a csv file of the taus file, need to be specified in case a causal hazard ratio under the inverse Gaussian need to be calculated}
}
\value{
Returns a vector the bootstrap sample causal hazard ratio estimator.
}
\description{
Calculation of the kernel-based and the Cox-based estimators for the causal hazard ratio for each bootstrap sample
}
\details{
This function calculate the kernel-based and the Cox-based estimators for the causal hazard ratio for each bootstrap sample. T
 This function can be used for RCT or for observational studies. For the observational studies setting, a vector name "w",of the observations' weights needs to pre-specified and need to be of the columns' .
 The weights are used also in the kernel-based estimator and ahe Cox-based estimator.
 This function calls \code{\link{muhaz_hr_for_real_data_example}} function that estimate the kernel-based casual hazard ratio in each sample,  \code{\link{coxph}} and \code{\link{base.haz.estimator}} ( from the survival package) to calculate the Cox-based estimator.
}
