% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muhaz_har_for_real_data_example.R
\name{muhaz_hr_for_real_data_example}
\alias{muhaz_hr_for_real_data_example}
\title{Title}
\usage{
muhaz_hr_for_real_data_example(
  frailty.type = 1,
  data,
  theta,
  confounder,
  n.est.grid = 51,
  b.cor = "n",
  end.time = NA,
  min.time = NA,
  kern = "e"
)
}
\arguments{
\item{frailty.type}{}

\item{data}{}

\item{theta}{}

\item{confounder}{}

\item{n.est.grid}{}

\item{b.cor}{}

\item{end.time}{}

\item{min.time}{}

\item{kern}{}
}
\value{

}
\description{
Title
}
