\name{print.pehaz}
\alias{print.pehaz}
\title{ Print a pehaz object. }
\description{
  The pehaz object is printed.
}
\usage{
\method{print}{pehaz}(x, \dots)
}

\arguments{
  \item{x}{ An object of class \code{pehaz}.  }
  \item{\dots}{ Ignored. }
}

\value{
 No value is returned. \code{x} is printed.
}
\seealso{ \code{\link{pehaz.object}}}

\keyword{survival}
