% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muhaz_har_for_real_data_example.R
\name{muhaz_hr_for_real_data_example}
\alias{muhaz_hr_for_real_data_example}
\title{Calculation of the kernel-based causal hazard ratio}
\usage{
muhaz_hr_for_real_data_example(
  frailty.type = 1,
  data,
  theta,
  confounder,
  n.est.grid = 51,
  b.cor = "n",
  end.time = NA,
  min.time = NA,
  kern = "e"
)
}
\arguments{
\item{frailty.type}{This function can calculate the causal hazard ratio under two frailty distribution-1=Gamma frailty distribution, 2=Inverse Gaussian distribution. The default value is the Gamma distribution.}

\item{data}{The data that the estimators will be estimated on. In case of RCT, a data with column named "time, treatment, status" needs to be specified. In case of observational studies a firth column named "W" that contains the weights need to be specified also.}

\item{theta}{The frailty variance}

\item{confounder}{"no"- for RCT setting, or in case a non-weighted estimator wanted to be calculated. "yes"-for observational studies setting, or in evey other case a weighted estimator need to be calculated.}

\item{n.est.grid}{Number of points in the estimation grid, where hazard estimates are computed. Default value is 101.}

\item{b.cor}{Boundary correction type. Possible values are: "none" - no boundary correction "left" - left only correction "both" - left and right corrections Default value is "both". Only the first letter needs to be given (e.g. b.cor="n").}

\item{end.time}{Right bound of the time domain used in analysis. If missing, max.time is the minumum between the (time at which ten patients remain at risk in the treatment group, time at which ten patients remain at risk in the control group)}

\item{min.time}{Left bound of the time domain used in analysis. If missing, max.time is the maximum between the (time at which the first event occur treatment group, time at which he first event occur  in the control group)}

\item{kern}{Boundary kernel function to be used. Possible values are: "rectangle", "epanechnikov", "biquadratic", "triquadratic". Default value is "epanechnikov". Only the first letter needs to be given (e.g. kern="b").}
}
\value{
Returs a data frame with the following columns
\itemize{
\item{time}{ -the estimatiom points}
\item{HR}{ -The kernel-baes causal hazard ratio}
\item{theta}{ -the selected frailty variance, theta}
\item{method}{- This is the estimation always "muhaza" }
}
}
\description{
This function calculate the kernel-based hazard ratio, by estimating first the kernel-based hazard rate and the cumulative hazard function for a given value of frailty variance, theta.
}
\details{
This function calculate the kernel-based hazard function for each tratment arm using the \code{\link{muhaz}} function. In the presence of confounders such the case of observational studies a weighted version of the hazard rate is calculated. 
 Its impotent to notice that in case of no-confounders three data coloumns needs to be specified: time, status and treatment. The possible values of the treatment needs to be :0,1.
 In case of confounders, a firth column named "W" that contains the observations weights need to be specified. After the calculation of the hazard rate, 
 the comulative hazard is calulated with \code{\link{integrate}} function.
}
